feat: implement waitlist functionality with form, services, and state management
