--- core/src/zap-old/services/waitlist/get-number-of-people-in-waitlist.service.ts
@@ -1,15 +1,0 @@
-import "server-only";
-
-import { getNumberOfPeopleInWaitlistQuery } from "@/zap/db/queries/waitlist.query";
-import { NotFoundError } from "@/zap/lib/api/errors";
-
-export async function getNumberOfPeopleInWaitlistService() {
-  const result = await getNumberOfPeopleInWaitlistQuery.execute();
-
-  if (!result.length) {
-    throw new NotFoundError("No waitlist records found");
-  }
-
-  const record = result[0];
-  return record.count;
-}
